requirements.txt
selenium
beautifulsoup4
webdriver-manager
faiss-cpu==1.10.0
fastembed==0.6.0
langchain==0.3.22
langchain-community==0.3.20
langchain-core==0.3.50
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.7
langchain_pinecone
pypdf==5.4.0
python-dotenv==1.1.0
pinecone~=6.0.2